requirements.txt
fastapi==0.104.1
pydantic==2.5.2
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
ultralytics==8.0.227